pyproject.toml
[tool.poetry]
name = "desafio-mb"
version = "0.1.0"
description = "MB"
authors = ["Diogo Lima"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.3"
environs = "^11.0.0"
requests = "^2.32.3"
uvicorn = "^0.32.0"
sqlmodel = "^0.0.22"
alembic = "^1.13.3"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
